Updating mainboard for nRF52 support.

+++ Nemesis-Mod/src/Pin.cpp
@@ -1,7 +1,7 @@
 #include <stddef.h>
 #include "Pin.h"
 
-Pin::Pin(uint8_t id, Mainboard* hal) {
+Pin::Pin(uint32_t id, Mainboard* hal) {
     m_Id = id;
     hardware = hal;
 }
